Refactor result card and hook file
