+++ src/components/Results/hooks.js
@@ -1,11 +1,10 @@
 import { useState, useEffect } from "react";
-import qs from "query-string";
+import { updateDigests, processQueryData } from "./helpers";
 import {
   mergeDigestsAndResult,
   getNotAnchoredDigests,
-  getDigests,
-  timestamp,
-  verify
+  handleVerify,
+  handleTimestamp
 } from "src/helpers/dcrtime";
 
 const useProcessDigests = location => {
@@ -18,8 +17,6 @@ const useProcessDigests = location => {
       try {
         setLoading(true);
         const verifyRes = await handleVerify(digests);
-        if (verifyRes.error) setError(new Error("Invalid array of digests"));
-
         const verifiedDigests = mergeDigestsAndResult(digests, verifyRes);
         setDigests(verifiedDigests);
 
@@ -36,7 +33,7 @@ const useProcessDigests = location => {
         ) {
           const tsRes = await handleTimestamp(notTimestampedDigests);
           const tsDigests = mergeDigestsAndResult(notTimestampedDigests, tsRes);
-          setDigests(digests => updateDigests(digests, tsDigests));
+          setDigests(updateDigests(digests, tsDigests));
         }
         setLoading(false);
       } catch (e) {
@@ -44,55 +41,12 @@ const useProcessDigests = location => {
       }
     }
 
-    // query string parameters
-    const {
-      digests: strDigests,
-      names: strNames,
-      timestamp: strTimestamp
-    } = qs.parse(location.hash);
-
-    const digests = strDigests.split(",");
-    const names = strNames ? strNames.split(",") : [];
-    const shouldTimestamp = strTimestamp === "true";
-
-    // file names are set from the query param if existent otherwise the
-    // digest is used as the name
-    const data = digests.map((d, i) => ({
-      name: names[i] || d,
-      digest: d
-    }));
+    const { data, shouldTimestamp } = processQueryData(location.hash);
 
     handleProcessDigests(data, shouldTimestamp);
   }, [location.hash]);
 
-  // Handlers for API calls
-  const handleVerify = async data => {
-    const digests = getDigests(data);
-    try {
-      const res = await verify(digests);
-      return res;
-    } catch (e) {
-      setError(e);
-    }
-  };
-
-  const handleTimestamp = async data => {
-    const digests = getDigests(data);
-    try {
-      const res = await timestamp(digests, "data");
-      return res;
-    } catch (e) {
-      setError(e);
-    }
-  };
-
   return { digests, loading, error };
 };
-
-const updateDigests = (digests, newDigests) =>
-  digests.map(d => {
-    const newDigest = newDigests.filter(nd => nd.digest === d.digest)[0];
-    return newDigest || d;
-  });
 
 export default useProcessDigests;
